💄Styles: プロフィール画面のコンポーネントにstateを定義
