--- src/features/users/Profile.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import ProfileForm from './ProfileForm';
-
-const Profile = () => {
-  return (
-    <>
-      <ProfileForm />
-    </>
-  );
-};
-
-export default Profile;
