+++ src/features/users/containers/Profile.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import avatar from '/src/assets/avatar.png';
+
+import ProfileForm from '../components/ProfileForm';
+
+const profile = {
+  nickname: '',
+  twitterUsername: '',
+  githubUsername: '',
+  description: '',
+  avatar,
+};
+
+const Profile = () => {
+  return (
+    <>
+      <ProfileForm {...profile} />
+    </>
+  );
+};
+
+export default Profile;
